WLBlog/requirements.txt
Django==2.1.7
PyMySQL==0.9.3
pytz==2018.9
Pillow==5.4.1
slugify==0.0.1
redis==3.2.1
pillow==5.4.1
django-haystack==2.8.1
Whoosh==2.7.4
jieba==0.39